refactor: organize main page components
