+++ app/(main)/communities/components/ManagedCommunitiesGrid.tsx
@@ -1,8 +1,8 @@
 import Link from 'next/link'
+import { copy } from '@/app/(main)/copy'
 import { Button } from '@/components/ui'
 import { Routes } from '@/lib/config'
 import type { RouterOutput } from '@/server/api'
-import { copy } from '../copy'
 import { EmptyState } from './EmptyState'
 import { GenericPagination } from './GenericPagination'
 import { ManagedCommunityCard } from './ManagedCommunityCard'
