+++ app/(main)/communities/components/ManagedCommunityCard.tsx
