+++ app/(main)/communities/components/UserCommunitiesList.tsx
@@ -1,5 +1,5 @@
+import { copy } from '@/app/(main)/copy'
 import type { RouterOutput } from '@/server/api'
-import { copy } from '../copy'
 import { EmptyState } from './EmptyState'
 import { GenericPagination } from './GenericPagination'
 import { UserCommunityItem } from './UserCommunityItem'
