+++ app/(main)/communities/components/UserCommunityItem.tsx
