+++ app/(main)/communities/components/index.ts
@@ -1,0 +1,9 @@
+export { CommunityHeader } from './CommunityHeader'
+export { CommunitySubscribeModal } from './CommunitySubscribeModal'
+export { CommunityTabs } from './CommunityTabs'
+export { ManagedCommunitiesGrid } from './ManagedCommunitiesGrid'
+export { ManagedCommunityCard } from './ManagedCommunityCard'
+export * from './ProgressiveCommunitiesList'
+export { ProgressiveCommunityView } from './ProgressiveCommunityView'
+export { UserCommunitiesList } from './UserCommunitiesList'
+export { UserCommunityItem } from './UserCommunityItem'
