+++ app/(main)/components/index.ts
@@ -1,34 +1,9 @@
-// Main Components
-export { CategoryDiscoverCard } from './CategoryDiscoverCard'
-export { CommunityDiscoverCard } from './CommunityDiscoverCard'
-export { CommunityHeader } from './CommunityHeader'
-export { CommunitySubscribeModal } from './CommunitySubscribeModal'
-export { CommunityTabs } from './CommunityTabs'
+export { ActiveLink } from './ActiveLink'
 export { EmptyState } from './EmptyState'
-export { EventCard } from './EventCard'
-export { EventDiscoverCard } from './EventDiscoverCard'
-export { EventForm } from './EventForm'
-export { EventPage } from './EventPage'
-export { EventRegisterModal } from './EventRegisterModal'
-export { EventsPagination } from './EventsPagination'
-export { FilteredEventsList } from './FilteredEventsList'
-export { LocationHeader } from './LocationHeader'
-export { LocationModal } from './LocationModal'
-export { LocationsDiscover } from './LocationsDiscover'
-export { ManagedCommunitiesGrid } from './ManagedCommunitiesGrid'
-export { ManagedCommunityCard } from './ManagedCommunityCard'
-export { ManageEventCard } from './ManageEventCard'
+export { GenericPagination } from './GenericPagination'
 export { Navbar } from './Navbar'
-export { NoEvents } from './NoEvents'
 export { PageWrapper } from './PageWrapper'
 export { PeriodTabs } from './PeriodTabs'
-export * from './ProgressiveCommunitiesList'
-export * from './ProgressiveEventPage'
-export * from './ProgressiveEventsList'
-export * from './ProgressiveManageEventCard'
-export { ResponsiveGridCarousel } from './ResponsiveGridCarousel'
 export { ShareActions } from './ShareActions'
 export { ShareLink } from './ShareLink'
-export { SimilarEvents } from './SimilarEvents'
-export { UserCommunitiesList } from './UserCommunitiesList'
-export { UserCommunityItem } from './UserCommunityItem'
+export { Stats } from './Stats'
