+++ app/(main)/events/components/CategoryDiscoverCard.tsx
