+++ app/(main)/events/components/CommunityDiscoverCard.tsx
