+++ app/(main)/events/components/EventCard.tsx
