+++ app/(main)/events/components/EventDiscoverCard.tsx
