+++ app/(main)/events/components/EventLocation.tsx
