+++ app/(main)/events/components/FilteredEventsList.tsx
