+++ app/(main)/events/components/ManageEventCard.tsx
