+++ app/(main)/events/components/ProgressiveEventPage.tsx
