+++ app/(main)/events/components/ProgressiveManageEventCard.tsx
