+++ app/(main)/events/components/ResponsiveGridCarousel.tsx
