+++ app/(main)/events/components/SimilarEvents.tsx
