+++ app/(main)/locations/components/LocationsDiscover.tsx
