+++ app/(main)/locations/components/index.ts
@@ -1,0 +1,3 @@
+export { LocationHeader } from './LocationHeader'
+export { LocationModal } from './LocationModal'
+export { LocationsDiscover } from './LocationsDiscover'
